Start allowing some spells to have offsets for where they originate from. Requires effectlib version change, the oldest version of effectlib usable is now 4.0
Add new config tool to the main spell class, getConfigVector which accepts a comma seperated vector in the form of x,y,z.

Fireball gets support for cast position offsets, aka the fireball can be launched from different locations relative to the caster. The config option for setting the relative offset is called relative-cast-position-offset and it accepts a the vector format. To use the relative offset, it is also required to set use-relative-cast-location-offset to true (it defaults to false for now so we avoid extra calculations). The absolute position offset is controlled by the config option absolute-cast-position-offset and also uses vector format. To enable the use of the absolute cast location offset, you must set use-absolute-cast-location-offset to true (defaults to false for now). Additionally, an option called do-offset-targeting-corrections was added to fireball and defaults to true. This will attempt to redirect any offset fireballs to target what the player was pointing at at cast time.