Added new condition worldguardmembership. The condition var may be set to owner, which will make the modifier require the player to be an owner of the region for the check to pass, otherwise, simple membership is sufficient.
Added new condition worldguardbooleanflag. The condition var should be the name of one of the worldguard flags as specified in the DefaultFlag class in worldguard. The condition var determines which flag to check for the highest priority region that the player is currently in. If the choosen flag is set to true in that region, the modifier check passes.

Added a new condition worldguardstateflag. The condition var should be the name of the worldguard flag to check as it is called in the DefaultFlag class of worldguard. If the flag's value is allow in the top priority region applicable to the player, then the modifier check passes.

The names for the flags can be found here https://github.com/sk89q/WorldGuard/blob/master/worldguard-legacy/src/main/java/com/sk89q/worldguard/protection/flags/DefaultFlag.java