Added modifiers: -moneymorethan -moneylessthan
Modified the /c modifyvariable command again.
This change allows you to modify a player's variable based upon the value of another player's variable.
The old syntax was
 /c modifyvariable <var> <player> [+|-|=|*|/]<amount or other variable name>

The new syntax is
/c modifyvariable <var> <player> [+|-|=|*|/]<amount|[target variable owner:]variable_name>
specifying target variable owner will have the operation use the value of that player's variable as <amount> instead of using <player>'s value of the variable. Note that this will not likely work yet with offline players.