Volatile code handlers now should be able to update textures of skull meta. Add a new option to predefined items called texture, accepts a string with the texture data. A called signature is also required to be defined for predefined items that specify a texture attribute.
Volatile code handlers now also get to set player skins.

Update the readme file with more up to date links for the resources.

Added "tags" as a new option to all spells. It accepts a list of strings as the value. This will be used in upcoming modifiers and checks to serve as simple metadata.

Created a spell filter class which accepts a list of spells to guarantee to pass, a list of spells to block, which is considered after the whitelist, a list of tags to look for, and a list of tags to avoid. The list of tags to avoid takes priority over the list of tags to look for. The explicit spell lists take priority over the tag lists. One tag matching is considered sufficient to meet the criteria for the allowed tags list and the disallowed tags list.

If no collections are specified, then all non null spells will be accepted.

Using the spell filter for the following spells:
-EmpowerSpell
-ClaritySpell
-SpellHasteSpell

The option for the explitly allowed spell list is "spells"
The option for the explitily denied spell list is "denied-spells"
The option for the allowed tags list is "spell-tags"
The option for the denied tags list is "denied-spell-tags"

Added OffsetLocationSpell, which extends targeted spell. It functions similarly to the LocationSpell, which allows you to set a fixed location to cast from, but instead of using a fixed location, it instead casts the subspell at a position offset.

The OffsetLocationSpell specifies the following config options:
- relative-offset which is of type vector
- absolute-offset which is of type vector
- spell which expects subspell format and must be a targeted location spell.

Note that the texture related options have NOT yet been tested and are very possible to fail.

Added SkinSpell which can set the skin that someone appears to have.
Option texture accepts type string
Option signature accepts type string