Added targeted.AgeSpell Implements TargetedEntitySpell
Has an option called "apply-age-lock" which defaults to false. If set to true, it will lock the target's age.
Has an option called "set-maturity" which defaults to true. When true, this spell will be allowed to modify the target's age.
Has an option called "age" which accepts an integer. This value will be the new age value for the target if "set-maturity" is true.