Only try setting the skulls the dangerous way if it is necessary (yes, I am calling using craftbukkit calls and reflection dangerous).
Added TargetingCondition which is referencable as "targeting".
The condition var can is optional and can be "caster" or a comma seperated list of entity types.
If there is no condition var, then the condition passes if the target is targeting something.
If the condition var is set to "caster", then the condition passes if the target is targeting the caster.
Otherwise, the condition passes if the entity type of the target's target is contained in the list of entity types specified by the condition var.