Added PowerLessThanCondition which is referencable as "powerlessthan". It accepts a float as the condition var and passes if the spell power is less than the condition var.
Added PowerGreaterThanCondition which is referencable as "powergreaterthan".
It accepts a float as the condition var and passes if the spell power is greater than the condition var.

Added PowerEqualsCondition which is referencable as "powerequals".
It accepts a float as the condition var and passes if the spell power equals the condition var.