Added utility method to MagicSpells.java for handling argument substitution.
Small fix for targeting in volatile code.

Added the beginning of argument substituion in spells, the format is
%arg:(index):defaultValue%
default value may not contain spaces.

Moving message handlers over to the new system in order to allow arguments to be used in more places.

Added new options to WitherSkullSpell.

* An option called "fire" which accepts a boolean and defaults to false. When set to true, the wither skull impact should theoretically cause fires to start.
* An option called "skull-name" which accepts a string. When not null and not empty, this will be used as a custom name for the skull.

Note that the argument substitution is not yet complete, it does not yet work in all spells and it will be completed, this is being pushed so that you guys can play around with it a bit.