Added SpellTagCondition, referencable as spelltag. The condition variable is a string that should match the tag a spell is required to have in order to pass.
Added SpellBeneficialCondition, referencable as beneficial.
The condition does not use a condition variable and will pass if the spell being checked is flagged as beneficial.

Added ProxyCondition which will be used for accessing conditions supplied by optional addons such as MagicSpellsFactions.
Added conditions to the factions addon
- factions:powerlessthan
- factions:powergreaterthan
- factions:powerequals

Added new condition to the factions addon called factions:hasfaction

Fixed some warnings that were showing up when they wern't supposed to.

Adding special variables, which are predefined and cannot currently be manually modified. They automatically update to reflect the state of players.
Variables added so far:
- meta_location_x
- meta_location_y
- meta_location_z
- meta_location_pitch
- meta_location_yaw

Note that this has not yet been tested, please use caution when trying this version.