Added new special built in variables: - meta_saturation - meta_experience_level - meta_experience_points - meta_remaining_air - meta_fly_speed - meta_walk_speed - meta_food_level - meta_entity_id - meta_fire_ticks - meta_fall_distance - meta_players_online
Most of the special built in variables can also be modified just like regular variables now. For example, changing the x coordinate variable for a player will result in the player being teleported. There is not yet safety checking on ranges, so use caution.

Factions addon should now honor the friendly fire flag on factions, so it will ignore faction relations if the caster and target are in faction claims where friendly fire is enabled.

Added some in game commands for manipulating the teams plugin.

Added a new pass condition to the multi condition called XOR.
The XOR condition passes if exactly 1 of the checks passed. The condition stops running when one of the following happens:
- all of the checks have been run.
- more than 1 of the checks have passed