Fixed no such field exception in volatile code handles (thanks to those helping me out with the testing!)
Added an option to the homing missile spell called intermediate-special-effect-locations which accepts a non negative integer. If the value is greater than 0, it specifies how many locations between the previous missile location and the current one should have the particle effect displayed. For example, setting intermediate-special-effect-locations to 1 will cause the effects to be displayed half way between the previous location of the missile and the current location of the missile.