Prepare for targeting non living entities
Add a way for spells to enforce specific rules on what they are capable of handling as targets

Added spigot 1.11.2 to the build path

Added an option clled potioncolor to predefined item processing for potions. It accepts a hex string representing the color of the potion item in rgb format. Will cause errors on slightly older versions of spigot.

Refactoring item processing

Added resource pack spell to send a resource pack to players based upon a url so you aren't limited to the single one specified by the server.properties. The config option is url and the spell class is .targeted.ResourcePackSpell