Allow attribute modifiers to optionally specify what slot to operate on. The value to provide for the slot matches what is acceptable to use in vanilla nbt. The new format is - type amount operation slot
Also updated the default configs a little.

Added a new option to the conjure spells called pickup-delay, which defaults to 0 and accepts an integer to apply as the pickup delay to item entities produced by this spell.

Preliminary version of the entity select spell. Doesn't currently have any spells that can work with it yet.