Added a new passive trigger called swaphanditem. Currently does not require any trigger variables. Fix up priorities of passive triggers, correct some handlings of ignoring canceled events.
Added a new option to passive spells called cancel-default-action-when-cast-fails. When set to true, the default action of the trigger will still be canceled if the spell fails as long as the cast was issued.

Modify the event registration in the MagicSpells class to allow overriding the listener priorities.

Added variants of existing triggers to fire at different event priorities. The preexisting ones will fire at normal priority.
For having the trigger fire at a different priority, simply append the appropriate suffix to the trigger name.
For priority lowest, append _lowestpriority to the trigger name.
For priority low, append _lowpriority to the trigger name.
For priority high, append _highpriority to the trigger name.
For priority highest, append _highestpriority to the trigger name.
For priority monitor, append _monitorpriority to the trigger name.
No name change is needed for the default (normal) priority.

EXCEPTION: this feature was not implemented for the resourcepack trigger due to it not being applicable.

Has not gone through much testing yet, feedback is appreaciated.