Refactor variable construction to have a dedicated class handle selecting which type based upon the type parameter in the config. Provide a method that can be overridden by variable classes to load additional data from the config.
Added a new option to passive spells called require-cancelled-event which defaults to false. When set to true, the spell will not cast if the event is not cancelled.

Corrected the handling of triggers which care about the state of an event being cancelled.

WARNING: this may break spells which depended upon previously flawed behavior.

Added a new variable type which may be referenced by specifying the config field, type, to distancetolocation
There are additional configuration fields to specify for this variable type.
There is a field called target-location which accepts a string in the format of world,x,y,z and defaults to world,0,0,0.
There is a field called cross-world which accepts a boolean and defaults to false. When set to true, distance will still be reported instead of returning the default value.
There is a field called cross-world-distance-multiplier which accepts a double and defaults to 1.0. When a distance is being reported that is actually between multiple worlds, the coordinate difference will be multiplied by this value.

Just a side note that this variable cannot be manually modified nor will its values be saved (it can't be set as a permanent variable).

Added a variant of the distancetolocation variable type referencable as squareddistancetolocation. Instead of calculating the distance between the locations, it just gives you the squared distance. Use this instead of the distancetolocation type whenever possible as it is much faster and avoids using the square root function (a very intensive function).

Explicitly cast null in some places in the memory addon to avoid overloaded method ambiguity.

Giving the cripple spell a little attention. Added new options to it.

use-slowness-effect accepts a boolean and defaults to true. When set to false, the cripple spell will no longer apply a slowness effect to the target.

apply-portal-cooldown accepts a boolean and defaults to false. When set to true, this spell will set the portal cooldown field of the target, if the duration of the new cooldown is longer than the existing cooldown.

portal-cooldown-ticks accepts an integer and defaults to 100. This number will be used as the target's new portal cooldown, if it is greater than the target's current portal cooldown (and if apply-portal-cooldown is set to true).

STRING VARIABLES
That's right, String variables are finally coming to MagicSpells.
To use them, start by setting type to playerstring
To set your default string value, just use the field default-value

Note: it is highly recommended to back up your variable data before loading this. This update contains significant changes and new concepts. I don't expect it to be problematic, but the nature of the changes warrants some extra caution. That said, enjoy.