Added targeted.ConversationSpell
Currently, the only configuration option that it reads on top of those common to all targeted spells is an option called conversation.
The option called conversation must be a configuration section in ConversationFactory format, as described in MagicSpells Config Objects.md

MagicSpells Config Objects.md will be used to specify various formats that are likely to be used in many places, specifically for representing serialized forms of objects.

For more information on this update, please read the previously mentioned file as it will also give you a sense of what you can do with this update.