More cleanup - Extracted the MagicConfig filename filter to a reusable constant.
Added a new cast modifier condition called "spellcaststate".
This condition only works in the normal modifiers section listening to the spell cast event.
The condition variable may be any of the following
- NORMAL
- ON_COOLDOWN
- MISSING_REAGENTS
- CANT_CAST
- NO_MAGIC_ZONE
- WRONG_WORLD

Improved null safety within the ClaritySpell.

Added a new cast modifier condition called "pluginenabled". It accepts the name of a plugin as the condition variable.
It passes if the targeted plugin is enabled.