Further cleanup Added additional meta variables - meta_velocity_x is linked to the x component of the player's velocity - meta_velocity_y is linked to the y component of the player's velocity - meta_velocity_z is linked to the z component of the player's velocity - meta_max_air is linked to the maximum air the player may have - meta_no_damage_ticks is linked to the player's no damage ticks - meta_max_no_damage_ticks is linked to the player's maximum no damage ticks - meta_last_damage is linked to the record of the last damage that the player took
Optimized building of some error messages.
Optimized dance cast pattern parsing.

Added a new option to FirenovaSpell called spell.
The spell specified here is expected to be a targeted location spell and will be cast at each block that is placed by the nova animation, if the spell is defined.

MarkSpell enhancement:
- MarkSpell is now a TargetedLocationSpell so it can mark locations other than where the caster actually is.