Moving to Java 8 Have residence pulled from the ci instead of storing in the repo.
Introduce AdminTeachSpell, currently only runs from console. It functions like a teach spell but using permission names to determine what spells to teach. The command format is
/c <spellname> <playername> <permissionname> [permissionname...]
This will not work properly if the player logs out while this is in progress, this is a known issue.
This attempts to run as much of the logic as possible off of the main thread and be very lightweight in comparison to bulk permission checks (cc: pex users).

Added Metrics! I understand metrics can be a little touchy for some people, but I really do value the information it can provide me. It helps me better understand how MagicSpells is being used, which means I can prioritize my optimizations better. But it also will help keep me working on MagicSpells regularly. That said, if you're really against Metrics, disable it. But if you guys have ideas on what might be fun stats to include (number of spells you have? number of variables? etc. I'm open for ideas! Let's make this fun).