Added buff.ImpactRecordSpell Records spells that you get hit by into a string variable.
The string variable to save the spell into is specified by the option variable-name.

There is an option called record-cancelled which defaults to false.
If set to true, cancelled spells will also be recorded.

There is an option called filter which is a configuration section containing the following:
An optional field called spells which accepts a list of spell names to accept.
An optional field called denied-spells which accepts a list of spell names to block.
An optional field called spell-tags which accepts a list of spell tags to allow.
An optional field called denied-spell-tags which accepts a list of spell tags to block.