Add new conditions
HasItemPrecise, referenceable as hasitemprecise
This condition functions just like the hasitem condition, but requires a predefined item and matches all aspects other than item count.

WearingPrecise, referenceable as wearingprecise
This condition functions just like the wearing condition, but with the same differences listed above.

HoldingPrecise, referencable as holdingprecise
This is the picky equivalent of the holding condition.