Added 2 new options to buff spells:
spell-on-use-increment which accepts the internal name of a spell. If defined, this will be cast each time the use counter is incremented through expected methods. Technical note: it runs just before the usage is incremented.

spell-on-cost which accepts the internal name of a spell. If defined, it will cast just before the usage cost charge is processed, even if there is nothing set for the usage cost itself.

Other Changes:
Cleanup, minor optimizations.
Started building with NoCheatPlus.
Introduced framework of exemption assistants (not yet being used, still wip).
Fixed the VariableCastSpell not applying formatting and variables to str-doesnt-contain-spell.
Precompiled coordinate matching pattern for GateSpell to improve performance.