Avoid using entity.getLocation as much since it makes use of defensive copying Use utility from AccessibleObject for reflection to reduce necessary checks.
Give ReflectionHelper some new tools.

Change the area specification options on the AreaEffectSpell to accept doubles instead of ints.
Changed max-distance in PulserSpell to accept a double instead of an integer for better control.
Changed retarget-reange and target-range in SpawnMonsterSpell to accept doubles instead of ints for better control.
Changed radius in ForcebombSpell to accept a double for better control.
Changed bounce-range in ChainSpell to a double for better control.
Changed range in RoarSpell to a double for better control.
Changed max-range in RecallSpell to a double for better control.
Changed range in PurgeSpell to a double for better control.

[WIP] Started a framework for command argument processing.
[WIP] template for reset spell

CRITICAL
introduced new option for the general config: ignore-grant-perms-fake-value
This defaults to true for avoiding changes in behavior. However, it is HIGHLY recommended to set this to false.