Added targeted.EntitySilenceSpell This spell controls the "silent" flag on living entities which may suppress sounds they usually make. There is an option called target-state which accepts a target boolean state and defaults to toggle.
Added targeted.SlimeSizeSpell
This spell is for changing the size of slimes.
It accepts an option called size which accepts a variable mod format (operation along with integer or variable name)
It accepts an option called min-size which accepts an integer and the minimum allowed value is 0. This is the minimum value to set a slime to.
It accepts an integer called max-size which defaults to 20 and is the maximum size a slime may be set to. If the maximum size is smaller than the minimum size, it is set to be equal to the minimum size.