Adding instant.BlockBeamSpell from ChronoKeeper! Produces a beam of blocks supported by armor stands, functions similarly to the beam spell.
Accepts an option called block-type which defaults to iron_block and sets what type of block the beam should be made of.

Accepts an option called max-distance of type float which defaults to 20.
This controls how far the beam may reach.

Accepts an option called hit-radius of type float which defaults to 1.
This controls the range from the center of the beam line from which an entity may be hit.

Accepts an option called remove-delay of type integer, which is the number of ticks before an armor stand should be removed.

Accepts an option called interval of type float which defaults to 1.
This controls how far apart the armor stands are.

Accepts an option called y-offset of type float which defaults to 0.5
This offsets the starting location of the beam.

Accepts an option called rotation-x of type float which defaults to 0.
This offsets the x rotation of the blocks displayed in the beam.

Accepts an option called rotation-y of type float which defaults to 0.
This offsets the y rotation of the blocks displayed in the beam.

Accepts an option called small of type boolean which defaults to false.
When set to true, the armor stands used will be smaller.

Accepts an option called spell of type string which defaults to an empty string. This option controls what spell the beam should cast on its targets.

Accepts an option called use-hp-fix of type boolean which defaults to false. Setting this to true sets the armor stand's hp to a large value as a workaround for an issue relating to the beam being cast in the same location multiple times and armor stands disappearing, if this issue occurs for you, set this option to true.