New operator for VariableMod ^ may now be used in variable mods for the power function. Add Shadow plugin for gradle
Add commons math as dependency.
Shade commons in.
Filter out non java resources.
Include apache commons in the compile group for the core jar.
Switch to optimized math functions in Util.

Note: this jar is a little large, I will work on trimming out unncessary files, this is just what the shade plugin is giving me to work with at the moment.
(Also, now is a good time to suggest new variable operations before the trimming)