Introduced the targeted.ParseSpell
The parse spell allows expected variables to be transferred into other variables.

While this was mainly made to bridge string variables and modifiers, It has much more potential uses than that and will most likely be expanded in the future!

Also added a header to Spells-Regular.yml to help newcomers get integrated into the community.