UnderBlock/OverBlock, OnLeash, VariableStringEquals Modifiers Added
UnderBlock | OverBlock
Example - underblock/overblock grass,3;20 required
Searches 20 blocks over/under an entity for a glass or block id of 3.

OnLeash
Example - onleash required
Checks if the target entity is on a leash.

VariableStringEquals
Example - variablestringequals class:archer required
Checks if the variable "class" is archer for the player.

- Added Line Breaks at EOF for respective files
- Updated UnderBlock Logic for target entities and players
- Used a Better Name for Variable String Comparison


Fix