Mob Utilities: Shear and Regrow Spell
Able to shear and regrow the wool of a sheep. Additional options for forcing the color of the sheep are available.

ShearSpell Example
    spell-class: ".targeted.ShearSpell"
    force-wool-color: false
    wool-count: 1-3
    #wool-color: Sheep color by default

wool-count: Either set a static amount of wool to be sheared or input a range for a count to be randomly chosen

force-wool-color: Will always force the color of the wool from the sheep to be specific dye even if the sheep is differently colored

wool-color: Sets the color of the forced wool when shearing the sheep; Must be used in conjuction with force-wool-color: true

===============

Regrow Spell Example
    spell-class: ".targeted.RegrowSpell"
    #wool-color: sheep's color by default
    force-wool-color: true

Same options as the shear spell but you can force the sheep to regrow a different color compared to it's original.

**If at any point an invalid wool-color is defined. The spells will always use the sheep's wool color regardless.**