Actually move all volatile code to their own classes
This should resolve startup errors for people trying to run MS on 1.14 or any version afterward even if there isn't a volatile code handler. With this being said, some features will not work properly without a volatile code handler due to the code not being in the Bukkit API, but the plugin should no longer error out on startup and at least function for the most part.